# api/swagger/swagger.yaml
definitions:
...
  question.createRequest:
    properties:
      content:
        additionalProperties: true
        description: Question content (structured)
        type: object
      created_by:
        description: Creator user ID
        type: string
      is_free:
        description: Access flag
        type: boolean
      level:
        description: Difficulty level
        type: string
      slug:
        description: Unique slug
        type: string
      tag_ids:
        description: Tags
        items:
          type: string
        type: array
      title:
        description: Question title
        type: string
      topic_id:
        description: Optional topic
        type: string
    required:
    - content
    - level
    - slug
    - title
    type: object
  question.createResponse:
    properties:
      id:
        description: New question ID
        type: string
    type: object
  question.listCardResponse:
    properties:
      id:
        type: string
      is_free:
        type: boolean
      is_new:
        type: boolean
      level:
        type: string
      slug:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  question.questionResponse:
    properties:
      content:
        additionalProperties: true
        type: object
      created_by:
        type: string
      id:
        type: string
      is_free:
        type: boolean
      level:
        type: string
      slug:
        type: string
      tag_ids:
        items:
          type: string
        type: array
      title:
        type: string
      topic_id:
        type: string
      updated_by:
        type: string
    type: object
  question.updateRequest:
    properties:
      content:
        additionalProperties: true
        description: Optional new content
        type: object
      is_free:
        description: Optional access flag
        type: boolean
      level:
        description: Optional new level
        type: string
      slug:
        description: Optional new slug
        type: string
      tag_ids:
        description: Optional new tags
        items:
          type: string
        type: array
      title:
        description: Optional new title
        type: string
      topic_id:
        description: Optional topic
        type: string
      updated_by:
        description: Updater user ID
        type: string
    type: object
  render.Error:
    properties:
      details: {}
      error:
        type: string
    type: object
...
  tag.tagResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /admin/questions:
    post:
      consumes:
      - application/json
      description: Create a new question
      parameters:
      - description: Question payload
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/question.createRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/question.createResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/render.Error'
        "409":
          description: Question already exists
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: Create question
      tags:
      - Questions
  /admin/questions/{id}:
    get:
      description: Returns full question including content (admin endpoint)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/question.questionResponse'
        "400":
          description: Invalid question ID
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Question not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: Get question by ID
      tags:
      - Questions
    put:
      consumes:
      - application/json
      description: Update an existing question
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Question update payload
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/question.updateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Question not found
          schema:
            $ref: '#/definitions/render.Error'
        "409":
          description: Question already exists
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: Update question
      tags:
      - Questions
...
  /questions:
    get:
      description: Returns paginated list of question cards with filters
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: Question level (beginner, medium, advanced)
        in: query
        name: level
        type: string
      - description: Comma separated tag slugs
        in: query
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/question.listCardResponse'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: List question cards
      tags:
      - Questions
  /questions/{slug}:
    get:
      description: Retrieve a question using its slug
      parameters:
      - description: Question slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/question.questionResponse'
        "400":
          description: Invalid slug
          schema:
            $ref: '#/definitions/render.Error'
        "404":
          description: Question not found
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: Get question by slug
      tags:
      - Questions
  /tags:
    get:
      description: Returns all tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/tag.tagResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: List tags
      tags:
      - Tags
...
  /topics/{id}/questions:
    get:
      description: Returns all questions for a specific topic
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/question.questionResponse'
            type: array
        "400":
          description: Invalid topic ID
          schema:
            $ref: '#/definitions/render.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/render.Error'
      summary: List questions by topic
      tags:
      - Questions
...
swagger: "2.0"